Add GET & POST Tests for Job Offer API
