+++ resources/js/components/ProjectsList.vue
@@ -110,7 +110,7 @@
                     <p> *Campos requeridos </p>
                 </h6>
                 <br>
-                <input class="button-1" type="submit" @click="update(project)" value="Actualizar">
+                <input class="button-1" type="button" @click="update(project)" value="Actualizar">
             </form>
         </pop-up>
